add mongo to mongo tests
