+++ ape-dts/src/test/README.md
@@ -1,19 +1,21 @@
 # setup test envrionments
 
 ## postgres
-- source
+
+### source
+
 ```
 docker run --name some-postgres-1 \
 -p 5433:5432 \
 -e POSTGRES_PASSWORD=postgres \
 -e TZ=Etc/GMT-8 \
--d postgis/postgis:latest 
+-d postgis/postgis:latest
 
 run `ALTER SYSTEM SET wal_level = logical;` and restart
 ```
 
+### target
 
-- target
 ```
 docker run --name some-postgres-2 \
 -p 5434:5432 \
@@ -23,6 +25,7 @@ docker run --name some-postgres-2 \
 ```
 
 - create a test db for EUC_CN in both source and target
+
 ```
 CREATE DATABASE postgres_euc_cn
   ENCODING 'EUC_CN'
@@ -31,9 +34,10 @@ CREATE DATABASE postgres_euc_cn
   TEMPLATE template0;
 ```
 
-
 ## mysql
-- source
+
+### source
+
 ```
 docker run -d --name some-mysql-1 \
 --platform linux/x86_64 \
@@ -51,7 +55,8 @@ docker run -d --name some-mysql-1 \
  --default_time_zone=+08:00
 ```
 
-- target
+### target
+
 ```
 docker run -d --name some-mysql-2 \
 --platform linux/x86_64 \
@@ -70,4 +75,61 @@ docker run -d --name some-mysql-2 \
  --binlog_rows_query_log_events=ON \
  --default_authentication_plugin=mysql_native_password \
  --default_time_zone="+07:00"
+```
+
+## mongo
+
+### source
+
+- 1. Create a Docker network:
+
+```
+docker network create mongo-network
+```
+
+- 2. Start 3 MongoDB containers for the replica set members:
+
+```
+docker run -d --name mongo1 --network mongo-network -p 9042:9042 mongo --replSet rs0  --port 9042
+docker run -d --name mongo2 --network mongo-network -p 9142:9142 mongo --replSet rs0  --port 9142
+docker run -d --name mongo3 --network mongo-network -p 9242:9242 mongo --replSet rs0  --port 9242
+```
+
+- 3. Connect to the primary (mongo1) and initialize the replica set:
+
+```
+docker exec -it mongo1 bash
+mongosh --host localhost --port 9042
+> config = {"_id" : "rs0", "members" : [{"_id" : 0,"host" : "mongo1:9042"},{"_id" : 1,"host" : "mongo2:9142"},{"_id" : 2,"host" : "mongo3:9242"}]}
+> rs.initiate(config)
+> rs.status()
+```
+
+- 4. Enable authentication on the admin database:
+
+```
+> use admin
+> db.createUser({user: "ape_dts", pwd: "123456", roles: ["root"]})
+```
+
+- 5. Update /etc/hosts on Mac
+
+```
+127.0.0.1 mongo1 mongo2 mongo3
+```
+
+- 6. Test the connection from Mac
+
+```
+mongo "mongodb://ape_dts:123456@mongo1:9042/?replicaSet=rs0"
+```
+
+### target
+
+```
+docker run -d --name dst-mongo \
+	-e MONGO_INITDB_ROOT_USERNAME=ape_dts \
+	-e MONGO_INITDB_ROOT_PASSWORD=123456 \
+  -p 27113:27017 \
+	mongo
 ```
