+++ ape-dts/src/test/mod.rs
@@ -1,8 +1,9 @@
 mod log_reader;
+mod mongo_to_mongo;
 mod mysql_to_foxlake;
 mod mysql_to_mysql;
 mod pg_to_pg;
 mod rdb_to_kafka;
 mod test_base;
 mod test_config_util;
-mod test_runner;
+mod test_runners;
