+++ ape-dts/src/test/mongo_to_mongo/cdc_basic_test/task_config.ini
@@ -2,7 +2,7 @@
 db_type=mongo
 extract_type=cdc
 resume_token=
-url=mongodb://ape_dts:123456@mongo-0.mongo:27017/?replicaSet=rs
+url=mongodb://ape_dts:123456@mongo1:9042/?replicaSet=rs0
 
 [filter]
 ignore_dbs=
@@ -15,7 +15,7 @@ do_events=insert,update,delete
 db_type=mongo
 sink_type=write
 batch_size=2
-url=mongodb://ape_dts:123456@localhost:27050
+url=mongodb://ape_dts:123456@localhost:27113
 
 [router]
 tb_map=
