+++ ape-dts/src/test/mongo_to_mongo/mod.rs
@@ -1,0 +1,2 @@
+mod cdc_tests;
+mod snapshot_tests;
