+++ ape-dts/src/test/mongo_to_mongo/snapshot_basic_test/dst_ddl.sql
@@ -1,0 +1,7 @@
+use test_db_1
+
+db.tb_1.drop()
+db.tb_2.drop()
+
+db.createCollection("tb_1");
+db.createCollection("tb_2");
