+++ ape-dts/src/test/mongo_to_mongo/snapshot_basic_test/task_config.ini
@@ -1,12 +1,12 @@
 [extractor]
 db_type=mongo
 extract_type=snapshot
-url=mongodb://ape_dts:123456@mongo-0.mongo:27017/?replicaSet=rs
+url=mongodb://ape_dts:123456@mongo1:9042/?replicaSet=rs0
 
 [sinker]
 db_type=mongo
 sink_type=write
-url=mongodb://ape_dts:123456@localhost:27050
+url=mongodb://ape_dts:123456@localhost:27113
 batch_size=2
 
 [filter]
